+++ src/main/java/com/example/demo/src/bookmark/BookmarkController.java
@@ -50,7 +50,7 @@ public class BookmarkController {
     @GetMapping("/bookmark")
     public BaseResponse<GetBookmarkAllRes> getBookmarkAll() {
         try {
-            long userIdx = jwtService.getUserIdx();
+            int userIdx = jwtService.getUserIdx();
             GetBookmarkAllRes getBookmarkAllRes = bookmarkProvider.getBookmarkAll(userIdx);
 
             return new BaseResponse<>(getBookmarkAllRes);
@@ -63,7 +63,7 @@ public class BookmarkController {
     @GetMapping("/bookmark/board")
     public BaseResponse<List<GetBookmarkBoardRes>> getBookmarkBoard() {
         try {
-            long userIdx = jwtService.getUserIdx();
+            int userIdx = jwtService.getUserIdx();
 
             List<GetBookmarkBoardRes> getBookmarkBoardRes = bookmarkProvider.getBookmarkBoard(userIdx);
             return new BaseResponse<>(getBookmarkBoardRes);
@@ -76,7 +76,7 @@ public class BookmarkController {
     @GetMapping("/bookmark/scholarship")
     public BaseResponse<List<GetBookmarkScholarshipRes>> getBookmarkScholarship() {
         try {
-            long userIdx = jwtService.getUserIdx();
+            int userIdx = jwtService.getUserIdx();
 
             List<GetBookmarkScholarshipRes> getBookmarkScholarshipRes = bookmarkProvider.getBookmarkScholarship(userIdx);
             return new BaseResponse<>(getBookmarkScholarshipRes);
@@ -89,7 +89,7 @@ public class BookmarkController {
     @GetMapping("/bookmark/support")
     public BaseResponse<List<GetBookmarkSupportRes>> getBookmarkSupport() {
         try {
-            long userIdx = jwtService.getUserIdx();
+            int userIdx = jwtService.getUserIdx();
 
             List<GetBookmarkSupportRes> getBookmarkSupportRes = bookmarkProvider.getBookmarkSupport(userIdx);
             return new BaseResponse<>(getBookmarkSupportRes);
@@ -99,9 +99,9 @@ public class BookmarkController {
     }
 
     @PostMapping("/scholarships/{scholarshipIdx}/bookmark")
-    public BaseResponse<String> postBookmarkScholarship(@PathVariable("scholarshipIdx") long scholarshipIdx) throws BaseException {
+    public BaseResponse<String> postBookmarkScholarship(@PathVariable("scholarshipIdx") int scholarshipIdx) throws BaseException {
         try {
-            long userIdx = jwtService.getUserIdx();
+            int userIdx = jwtService.getUserIdx();
             PostBookmarkScholarshipReq postBookmarkScholarshipReq = new PostBookmarkScholarshipReq(userIdx, scholarshipIdx);
             bookmarkService.postBookmarkScholarship(postBookmarkScholarshipReq);
 
@@ -112,9 +112,9 @@ public class BookmarkController {
     }
 
     @PostMapping("/supports/{supportIdx}/bookmark")
-    public BaseResponse<String> postBookmarkSupport(@PathVariable("supportIdx") long supportIdx) throws BaseException {
+    public BaseResponse<String> postBookmarkSupport(@PathVariable("supportIdx") int supportIdx) throws BaseException {
         try {
-            long userIdx = jwtService.getUserIdx();
+            int userIdx = jwtService.getUserIdx();
             PostBookmarkSupportReq postBookmarkSupportReq = new PostBookmarkSupportReq(userIdx, supportIdx);
 
             bookmarkService.postBookmarkSupport(postBookmarkSupportReq);
